Fix ecef transform for projection estimate
Fix the bias introduced by the ECEF transform when computing the
reprojection error of tie points in bil sequences.

Also, temporarily disable using the initial trajectory, it seems that
after the ECEF correction, things match the expectation quite well.
(This could be re-introduced in the future as an option, but for the
time being, it just makes the computations slow as heck...)